refactor(layout,#171): complete Phase B – generalized manifest validation rule & domain LayoutEngine
- Generalize require-manifest-validation ESLint rule to support domain path and ignore pure re-export shims
- Add expanded rule tests (legacy path, domain path, shim exemption)
- Relocate LayoutEngine implementation to src/domain/layout; root file now a transitional re-export
- All tests green (169 passed, 1 skipped) and build succeeds
- ADR updated: status Active, Phase B (B1/B2) marked complete
